requirements.txt
sounddevice==0.5.2
numpy==1.24.3
scipy==1.11.1
speechrecognition==3.10.0
pyannote.audio==3.1.1
torch==2.0.1
torchaudio==2.0.2
transformers==4.30.2
openai==1.30.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio==3.4.3
websockets==11.0.3
aiofiles==24.1.0